v1

+++ Charly.c
@@ -1,6 +1,6 @@
-#include<stdlib.h>
 
-double Taxes(double value){
+
+double Taxes(float value){
     if(value>=15000 && value<50000){
         return 0.30*value;
     }else if(value>=50000)return 0.50*value;
